reorganize framework builder modules
